requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
Django==1.11.2
django-bootstrap3==8.2.3
django-haystack==2.6.1
django-mptt==0.8.7
elasticsearch==5.4.0
kombu==4.0.2
psycopg2==2.7.1
pyelasticsearch==1.4
pytz==2017.2
simplejson==3.11.0
six==1.10.0
urllib3==1.21.1
vine==1.1.3
